UPDATE [15/02]
- FIXED : Test for dataset class computation
	  Everything is now doubles (for now)
   	  Columns are now one vector instead of splitted between lines
           shall be faster to copy at some point
- TO DO :
 	  Need to make tests for the computation part
	  Fix build

+++ include/PredictionModel/BaggingModel.hpp
@@ -41,7 +41,7 @@ public:
   void aggregate_Forest(const std::map<int, DecisionTree> &forest);
   void train(const DataSet &data);
   void train(const DataSet &data, int n);
-  std::vector<float> predict(const DataSet &datas);
+  std::vector<double> predict(const DataSet &datas);
 };
 
 #endif
